Clean up ADBC implementation to align with existing IDatabricksClient integration
This commit removes unnecessary FlightSQL integration and implements proper
Arrow IPC streaming that works with the existing Databricks server infrastructure.

## Changes:

### 1. Removed Unnecessary FlightSQL Code
- Deleted `AdbcFlightSqlIntegration.java` - not needed for IDatabricksClient integration
- Deleted `AdbcFlightStream.java` - FlightSQL approach doesn't align with existing architecture
- Focuses on working within existing Databricks server communication protocols

### 2. Implemented Proper ArrowIPC Integration
- Updated `ArrowIpcStreamIterator` to work with existing `ArrowStreamResult`
- Added actual implementation for `getArrowAllocator()` using RootAllocator
- Implemented `getArrowSchemaFromResult()` using existing VectorSchemaRoot iterator
- Removed TODO placeholders with working implementations

### 3. Enhanced AdbcDatabricksResultSet
- Proper schema extraction from existing Arrow infrastructure
- BufferAllocator management aligned with Arrow patterns
- Integration with existing ArrowVectorSchemaRootIterator
- Works seamlessly with IDatabricksClient response handling

### 4. Simplified Architecture
- No external dependencies on FlightSQL endpoints
- Uses existing DatabricksResultSet -> ArrowStreamResult -> VectorSchemaRoot flow
- Arrow IPC serialization happens at the client side from existing data
- Maintains backward compatibility with all JDBC functionality

## Technical Benefits:

- **Server Compatibility**: Works with existing Databricks server endpoints
- **No Protocol Changes**: Uses current IDatabricksClient communication
- **Arrow Integration**: Leverages existing Arrow infrastructure
- **Resource Management**: Proper BufferAllocator lifecycle management
- **Performance**: Direct Arrow IPC serialization from existing VectorSchemaRoot

This implementation provides ADBC Arrow IPC streaming capability while
remaining fully compatible with the existing Databricks JDBC architecture.
