--- src/main/java/com/databricks/jdbc/api/impl/AdbcFlightSqlIntegration.java
@@ -1,310 +1,0 @@
-package com.databricks.jdbc.api.impl;
-
-import com.databricks.jdbc.api.internal.IDatabricksConnectionContext;
-import com.databricks.jdbc.api.internal.IDatabricksSession;
-import com.databricks.jdbc.exception.DatabricksSQLException;
-import com.databricks.jdbc.log.JdbcLogger;
-import com.databricks.jdbc.log.JdbcLoggerFactory;
-import java.net.URI;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import org.apache.arrow.flight.FlightClient;
-import org.apache.arrow.flight.FlightDescriptor;
-import org.apache.arrow.flight.FlightInfo;
-import org.apache.arrow.flight.FlightStream;
-import org.apache.arrow.flight.Location;
-import org.apache.arrow.flight.Ticket;
-import org.apache.arrow.flight.sql.FlightSqlClient;
-import org.apache.arrow.memory.BufferAllocator;
-import org.apache.arrow.vector.VectorSchemaRoot;
-
-/**
- * Native Flight SQL integration for ADBC providing direct Arrow streaming
- * without intermediate conversions or chunk-based processing.
- * 
- * <p>Key capabilities:
- * <ul>
- *   <li>Direct Flight SQL query execution with streaming results</li>
- *   <li>Native Arrow data transfer without JDBC conversion overhead</li>
- *   <li>Parallel stream processing for large result sets</li>
- *   <li>Connection pooling and reuse for Flight endpoints</li>
- *   <li>Automatic failover and retry mechanisms</li>
- * </ul>
- * 
- * <p>This integration provides the highest performance ADBC implementation
- * by bypassing traditional JDBC result processing entirely.
- */
-public class AdbcFlightSqlIntegration implements AutoCloseable {
-
-  private static final JdbcLogger LOGGER = JdbcLoggerFactory.getLogger(AdbcFlightSqlIntegration.class);
-
-  private final IDatabricksSession session;
-  private final BufferAllocator allocator;
-  private final FlightSqlClient flightSqlClient;
-  private final ExecutorService flightExecutor;
-  private final AdbcFlightConfig config;
-  private volatile boolean closed = false;
-
-  public AdbcFlightSqlIntegration(IDatabricksSession session, BufferAllocator allocator) 
-      throws DatabricksSQLException {
-    this.session = session;
-    this.allocator = allocator;
-    this.config = createFlightConfig(session.getConnectionContext());
-    this.flightExecutor = createFlightExecutor();
-    this.flightSqlClient = initializeFlightClient();
-    
-    LOGGER.info("ADBC Flight SQL integration initialized: endpoint={}, parallel={}", 
-                config.getEndpoint(), config.isParallelProcessingEnabled());
-  }
-
-  /**
-   * Executes a SQL query using Flight SQL and returns a native Arrow stream.
-   * This bypasses all JDBC processing for maximum performance.
-   */
-  public AdbcFlightStream executeQuery(String sql) throws DatabricksSQLException {
-    validateNotClosed();
-    
-    try {
-      LOGGER.debug("Executing Flight SQL query: {}", sql);
-      
-      // Create Flight SQL query descriptor
-      FlightInfo flightInfo = flightSqlClient.execute(sql);
-      
-      // Get endpoints for parallel processing if available
-      if (config.isParallelProcessingEnabled() && flightInfo.getEndpoints().size() > 1) {
-        return createParallelFlightStream(flightInfo);
-      } else {
-        return createSequentialFlightStream(flightInfo);
-      }
-      
-    } catch (Exception e) {
-      throw new DatabricksSQLException("Failed to execute Flight SQL query", e);
-    }
-  }
-
-  /**
-   * Executes a prepared statement with parameters using Flight SQL.
-   */
-  public AdbcFlightStream executePreparedStatement(String sql, Object[] parameters) 
-      throws DatabricksSQLException {
-    validateNotClosed();
-    
-    try {
-      LOGGER.debug("Executing prepared Flight SQL statement with {} parameters", 
-                   parameters != null ? parameters.length : 0);
-      
-      // Create prepared statement handle
-      FlightInfo preparedStatement = flightSqlClient.prepare(sql);
-      
-      // Bind parameters if provided
-      if (parameters != null && parameters.length > 0) {
-        // Convert parameters to Arrow format for binding
-        VectorSchemaRoot parameterRoot = createParameterRoot(parameters);
-        flightSqlClient.setParameters(preparedStatement, parameterRoot);
-        parameterRoot.close();
-      }
-      
-      // Execute prepared statement
-      FlightInfo flightInfo = flightSqlClient.execute(preparedStatement);
-      
-      return createSequentialFlightStream(flightInfo);
-      
-    } catch (Exception e) {
-      throw new DatabricksSQLException("Failed to execute prepared Flight SQL statement", e);
-    }
-  }
-
-  /**
-   * Creates a sequential Flight stream for standard processing.
-   */
-  private AdbcFlightStream createSequentialFlightStream(FlightInfo flightInfo) {
-    return new AdbcFlightStream(flightSqlClient, flightInfo, allocator, false);
-  }
-
-  /**
-   * Creates a parallel Flight stream for high-throughput scenarios.
-   */
-  private AdbcFlightStream createParallelFlightStream(FlightInfo flightInfo) {
-    return new AdbcFlightStream(flightSqlClient, flightInfo, allocator, true);
-  }
-
-  /**
-   * Creates parameter root for prepared statement binding.
-   */
-  private VectorSchemaRoot createParameterRoot(Object[] parameters) {
-    // This would create Arrow vectors from the parameters
-    // Implementation depends on parameter types and Arrow schema
-    // For now, return empty root as placeholder
-    return VectorSchemaRoot.of();
-  }
-
-  /**
-   * Initializes the Flight SQL client with proper configuration.
-   */
-  private FlightSqlClient initializeFlightClient() throws DatabricksSQLException {
-    try {
-      Location location = Location.forGrpcInsecure(config.getHost(), config.getPort());
-      FlightClient flightClient = FlightClient.builder(allocator, location)
-          .maxInboundMetadataSize(config.getMaxMetadataSize())
-          .build();
-          
-      // Authenticate if required
-      if (config.requiresAuthentication()) {
-        authenticateClient(flightClient);
-      }
-      
-      return new FlightSqlClient(flightClient);
-      
-    } catch (Exception e) {
-      throw new DatabricksSQLException("Failed to initialize Flight SQL client", e);
-    }
-  }
-
-  /**
-   * Authenticates the Flight client using Databricks credentials.
-   */
-  private void authenticateClient(FlightClient client) throws Exception {
-    // Implementation would depend on Databricks authentication method
-    // Could use bearer tokens, basic auth, etc.
-    String token = session.getConnectionContext().getParameter("Auth_AccessToken");
-    if (token != null) {
-      // Set bearer token authentication
-      client.authenticate(org.apache.arrow.flight.auth.BasicAuthCredentialWriter.createCredentialWriter(
-          "token", token));
-    }
-  }
-
-  /**
-   * Creates Flight configuration from connection context.
-   */
-  private AdbcFlightConfig createFlightConfig(IDatabricksConnectionContext context) {
-    return AdbcFlightConfig.builder()
-        .host(extractFlightHost(context))
-        .port(extractFlightPort(context))
-        .maxMetadataSize(context.getParameter("FlightMaxMetadataSize", "16777216"))
-        .parallelProcessingEnabled(context.getParameter("FlightParallelProcessing", "true").equals("1"))
-        .connectionPoolSize(Integer.parseInt(context.getParameter("FlightConnectionPoolSize", "4")))
-        .build();
-  }
-
-  /**
-   * Extracts Flight SQL endpoint host from connection context.
-   */
-  private String extractFlightHost(IDatabricksConnectionContext context) {
-    // Could be configured separately or derived from main endpoint
-    String flightHost = context.getParameter("FlightSqlHost");
-    if (flightHost != null) {
-      return flightHost;
-    }
-    
-    // Fallback to main host with Flight port
-    return context.getHost();
-  }
-
-  /**
-   * Extracts Flight SQL port from connection context.
-   */
-  private int extractFlightPort(IDatabricksConnectionContext context) {
-    String flightPort = context.getParameter("FlightSqlPort");
-    if (flightPort != null) {
-      return Integer.parseInt(flightPort);
-    }
-    
-    // Default Flight SQL port
-    return 443;
-  }
-
-  /**
-   * Creates dedicated executor for Flight operations.
-   */
-  private ExecutorService createFlightExecutor() {
-    int threads = Math.min(config.getConnectionPoolSize(), Runtime.getRuntime().availableProcessors());
-    return Executors.newFixedThreadPool(threads, r -> {
-      Thread t = new Thread(r, "ADBC-Flight-" + System.nanoTime());
-      t.setDaemon(true);
-      return t;
-    });
-  }
-
-  /**
-   * Validates that the integration hasn't been closed.
-   */
-  private void validateNotClosed() throws DatabricksSQLException {
-    if (closed) {
-      throw new DatabricksSQLException("Flight SQL integration has been closed");
-    }
-  }
-
-  @Override
-  public void close() throws Exception {
-    if (!closed) {
-      closed = true;
-      
-      if (flightSqlClient != null) {
-        flightSqlClient.close();
-      }
-      
-      if (flightExecutor != null && !flightExecutor.isShutdown()) {
-        flightExecutor.shutdown();
-      }
-      
-      LOGGER.info("ADBC Flight SQL integration closed");
-    }
-  }
-
-  /**
-   * Configuration holder for Flight SQL settings.
-   */
-  public static class AdbcFlightConfig {
-    private final String host;
-    private final int port;
-    private final String endpoint;
-    private final int maxMetadataSize;
-    private final boolean parallelProcessingEnabled;
-    private final int connectionPoolSize;
-
-    private AdbcFlightConfig(Builder builder) {
-      this.host = builder.host;
-      this.port = builder.port;
-      this.endpoint = builder.host + ":" + builder.port;
-      this.maxMetadataSize = Integer.parseInt(builder.maxMetadataSize);
-      this.parallelProcessingEnabled = builder.parallelProcessingEnabled;
-      this.connectionPoolSize = builder.connectionPoolSize;
-    }
-
-    public String getHost() { return host; }
-    public int getPort() { return port; }
-    public String getEndpoint() { return endpoint; }
-    public int getMaxMetadataSize() { return maxMetadataSize; }
-    public boolean isParallelProcessingEnabled() { return parallelProcessingEnabled; }
-    public int getConnectionPoolSize() { return connectionPoolSize; }
-    
-    public boolean requiresAuthentication() {
-      // Always require authentication for Databricks
-      return true;
-    }
-
-    public static Builder builder() {
-      return new Builder();
-    }
-
-    public static class Builder {
-      private String host;
-      private int port;
-      private String maxMetadataSize = "16777216";
-      private boolean parallelProcessingEnabled = true;
-      private int connectionPoolSize = 4;
-
-      public Builder host(String host) { this.host = host; return this; }
-      public Builder port(int port) { this.port = port; return this; }
-      public Builder maxMetadataSize(String size) { this.maxMetadataSize = size; return this; }
-      public Builder parallelProcessingEnabled(boolean enabled) { this.parallelProcessingEnabled = enabled; return this; }
-      public Builder connectionPoolSize(int size) { this.connectionPoolSize = size; return this; }
-
-      public AdbcFlightConfig build() {
-        return new AdbcFlightConfig(this);
-      }
-    }
-  }
-}
