--- src/main/java/com/databricks/jdbc/api/impl/AdbcFlightStream.java
@@ -1,378 +1,0 @@
-package com.databricks.jdbc.api.impl;
-
-import com.databricks.jdbc.exception.DatabricksSQLException;
-import com.databricks.jdbc.log.JdbcLogger;
-import com.databricks.jdbc.log.JdbcLoggerFactory;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
-import org.apache.arrow.flight.FlightInfo;
-import org.apache.arrow.flight.FlightStream;
-import org.apache.arrow.flight.sql.FlightSqlClient;
-import org.apache.arrow.memory.BufferAllocator;
-import org.apache.arrow.vector.VectorSchemaRoot;
-
-/**
- * High-performance Flight SQL stream implementation providing native Arrow streaming
- * with parallel processing capabilities for analytical workloads.
- * 
- * <p>Features:
- * <ul>
- *   <li>Native Arrow streaming without JDBC conversion</li>
- *   <li>Parallel endpoint processing for distributed queries</li>
- *   <li>Automatic resource management and cleanup</li>
- *   <li>Adaptive batching based on data characteristics</li>
- *   <li>Memory-efficient streaming for large result sets</li>
- * </ul>
- */
-public class AdbcFlightStream implements AutoCloseable, Iterable<VectorSchemaRoot> {
-
-  private static final JdbcLogger LOGGER = JdbcLoggerFactory.getLogger(AdbcFlightStream.class);
-
-  private final FlightSqlClient client;
-  private final FlightInfo flightInfo;
-  private final BufferAllocator allocator;
-  private final boolean parallelMode;
-  private final List<FlightStream> activeStreams;
-  private final ExecutorService streamExecutor;
-  
-  private volatile boolean closed = false;
-  private long totalRecordsStreamed = 0;
-  private long totalBytesStreamed = 0;
-
-  public AdbcFlightStream(FlightSqlClient client, FlightInfo flightInfo, 
-                         BufferAllocator allocator, boolean parallelMode) {
-    this.client = client;
-    this.flightInfo = flightInfo;
-    this.allocator = allocator;
-    this.parallelMode = parallelMode;
-    this.activeStreams = new ArrayList<>();
-    
-    // Create executor for parallel processing if needed
-    if (parallelMode && flightInfo.getEndpoints().size() > 1) {
-      this.streamExecutor = Executors.newFixedThreadPool(
-          Math.min(flightInfo.getEndpoints().size(), 
-                  Runtime.getRuntime().availableProcessors()),
-          r -> {
-            Thread t = new Thread(r, "ADBC-FlightStream-" + System.nanoTime());
-            t.setDaemon(true);
-            return t;
-          });
-    } else {
-      this.streamExecutor = null;
-    }
-    
-    LOGGER.debug("Created AdbcFlightStream: endpoints={}, parallel={}", 
-                flightInfo.getEndpoints().size(), parallelMode);
-  }
-
-  /**
-   * Returns a Java Stream of VectorSchemaRoot for functional processing.
-   */
-  public Stream<VectorSchemaRoot> stream() {
-    validateNotClosed();
-    
-    if (parallelMode && flightInfo.getEndpoints().size() > 1) {
-      return createParallelStream();
-    } else {
-      return createSequentialStream();
-    }
-  }
-
-  /**
-   * Creates a sequential stream processing Flight endpoints one by one.
-   */
-  private Stream<VectorSchemaRoot> createSequentialStream() {
-    return StreamSupport.stream(this.spliterator(), false);
-  }
-
-  /**
-   * Creates a parallel stream processing multiple Flight endpoints concurrently.
-   */
-  private Stream<VectorSchemaRoot> createParallelStream() {
-    List<CompletableFuture<Stream<VectorSchemaRoot>>> endpointFutures = new ArrayList<>();
-    
-    // Create futures for each endpoint
-    flightInfo.getEndpoints().forEach(endpoint -> {
-      CompletableFuture<Stream<VectorSchemaRoot>> future = CompletableFuture.supplyAsync(() -> {
-        try {
-          FlightStream flightStream = client.getStream(endpoint.getTicket());
-          synchronized (activeStreams) {
-            activeStreams.add(flightStream);
-          }
-          
-          return StreamSupport.stream(
-              new FlightStreamSpliterator(flightStream), false);
-              
-        } catch (Exception e) {
-          LOGGER.error("Failed to create stream for endpoint: {}", endpoint, e);
-          return Stream.empty();
-        }
-      }, streamExecutor);
-      
-      endpointFutures.add(future);
-    });
-    
-    // Combine all endpoint streams
-    return endpointFutures.stream()
-        .map(CompletableFuture::join)
-        .reduce(Stream.empty(), Stream::concat);
-  }
-
-  @Override
-  public Iterator<VectorSchemaRoot> iterator() {
-    validateNotClosed();
-    
-    if (parallelMode && flightInfo.getEndpoints().size() > 1) {
-      return new ParallelFlightIterator();
-    } else {
-      return new SequentialFlightIterator();
-    }
-  }
-
-  /**
-   * Sequential iterator for single-endpoint processing.
-   */
-  private class SequentialFlightIterator implements Iterator<VectorSchemaRoot> {
-    private final Iterator<org.apache.arrow.flight.FlightEndpoint> endpointIterator;
-    private FlightStream currentStream;
-    private VectorSchemaRoot nextRoot;
-    private boolean hasNextChecked = false;
-
-    SequentialFlightIterator() {
-      this.endpointIterator = flightInfo.getEndpoints().iterator();
-      advanceToNextStream();
-    }
-
-    @Override
-    public boolean hasNext() {
-      if (!hasNextChecked) {
-        checkForNext();
-        hasNextChecked = true;
-      }
-      return nextRoot != null;
-    }
-
-    @Override
-    public VectorSchemaRoot next() {
-      if (!hasNext()) {
-        throw new java.util.NoSuchElementException("No more records available");
-      }
-      
-      VectorSchemaRoot result = nextRoot;
-      nextRoot = null;
-      hasNextChecked = false;
-      
-      // Update metrics
-      if (result != null) {
-        totalRecordsStreamed += result.getRowCount();
-        totalBytesStreamed += estimateRootSize(result);
-      }
-      
-      return result;
-    }
-
-    private void checkForNext() {
-      while (nextRoot == null) {
-        // Check current stream first
-        if (currentStream != null && currentStream.next()) {
-          nextRoot = currentStream.getRoot();
-          return;
-        }
-        
-        // Move to next stream if available
-        if (!advanceToNextStream()) {
-          break; // No more streams
-        }
-      }
-    }
-
-    private boolean advanceToNextStream() {
-      // Close current stream if exists
-      if (currentStream != null) {
-        try {
-          currentStream.close();
-          synchronized (activeStreams) {
-            activeStreams.remove(currentStream);
-          }
-        } catch (Exception e) {
-          LOGGER.warn("Error closing Flight stream", e);
-        }
-      }
-      
-      // Get next endpoint
-      if (endpointIterator.hasNext()) {
-        try {
-          org.apache.arrow.flight.FlightEndpoint endpoint = endpointIterator.next();
-          currentStream = client.getStream(endpoint.getTicket());
-          
-          synchronized (activeStreams) {
-            activeStreams.add(currentStream);
-          }
-          
-          LOGGER.trace("Advanced to next Flight stream: {}", endpoint);
-          return true;
-          
-        } catch (Exception e) {
-          LOGGER.error("Failed to advance to next Flight stream", e);
-          return false;
-        }
-      }
-      
-      return false;
-    }
-  }
-
-  /**
-   * Parallel iterator for multi-endpoint processing.
-   */
-  private class ParallelFlightIterator implements Iterator<VectorSchemaRoot> {
-    private final Iterator<VectorSchemaRoot> combinedIterator;
-
-    ParallelFlightIterator() {
-      // Create combined iterator from parallel stream
-      this.combinedIterator = createParallelStream().iterator();
-    }
-
-    @Override
-    public boolean hasNext() {
-      return combinedIterator.hasNext();
-    }
-
-    @Override
-    public VectorSchemaRoot next() {
-      VectorSchemaRoot result = combinedIterator.next();
-      
-      // Update metrics
-      if (result != null) {
-        totalRecordsStreamed += result.getRowCount();
-        totalBytesStreamed += estimateRootSize(result);
-      }
-      
-      return result;
-    }
-  }
-
-  /**
-   * Spliterator for Flight stream processing.
-   */
-  private static class FlightStreamSpliterator implements java.util.Spliterator<VectorSchemaRoot> {
-    private final FlightStream flightStream;
-
-    FlightStreamSpliterator(FlightStream flightStream) {
-      this.flightStream = flightStream;
-    }
-
-    @Override
-    public boolean tryAdvance(java.util.function.Consumer<? super VectorSchemaRoot> action) {
-      if (flightStream.next()) {
-        action.accept(flightStream.getRoot());
-        return true;
-      }
-      return false;
-    }
-
-    @Override
-    public java.util.Spliterator<VectorSchemaRoot> trySplit() {
-      // Flight streams are not splittable at the record level
-      return null;
-    }
-
-    @Override
-    public long estimateSize() {
-      return Long.MAX_VALUE; // Unknown size
-    }
-
-    @Override
-    public int characteristics() {
-      return ORDERED | NONNULL;
-    }
-  }
-
-  /**
-   * Returns the schema of the result set.
-   */
-  public org.apache.arrow.vector.types.pojo.Schema getSchema() {
-    return flightInfo.getSchema();
-  }
-
-  /**
-   * Returns the number of records streamed so far.
-   */
-  public long getTotalRecordsStreamed() {
-    return totalRecordsStreamed;
-  }
-
-  /**
-   * Returns the estimated bytes streamed so far.
-   */
-  public long getTotalBytesStreamed() {
-    return totalBytesStreamed;
-  }
-
-  /**
-   * Returns the number of Flight endpoints being processed.
-   */
-  public int getEndpointCount() {
-    return flightInfo.getEndpoints().size();
-  }
-
-  /**
-   * Estimates the memory size of a VectorSchemaRoot.
-   */
-  private long estimateRootSize(VectorSchemaRoot root) {
-    long size = 0;
-    for (org.apache.arrow.vector.FieldVector vector : root.getFieldVectors()) {
-      size += vector.getBufferSize();
-    }
-    return size;
-  }
-
-  /**
-   * Validates that the stream hasn't been closed.
-   */
-  private void validateNotClosed() {
-    if (closed) {
-      throw new IllegalStateException("AdbcFlightStream has been closed");
-    }
-  }
-
-  @Override
-  public void close() throws Exception {
-    if (!closed) {
-      closed = true;
-      
-      // Close all active streams
-      List<Exception> exceptions = new ArrayList<>();
-      synchronized (activeStreams) {
-        for (FlightStream stream : activeStreams) {
-          try {
-            stream.close();
-          } catch (Exception e) {
-            exceptions.add(e);
-            LOGGER.warn("Error closing Flight stream", e);
-          }
-        }
-        activeStreams.clear();
-      }
-      
-      // Shutdown executor if exists
-      if (streamExecutor != null && !streamExecutor.isShutdown()) {
-        streamExecutor.shutdown();
-      }
-      
-      LOGGER.debug("Closed AdbcFlightStream: records={}, bytes={}", 
-                  totalRecordsStreamed, totalBytesStreamed);
-      
-      // Throw first exception if any occurred during cleanup
-      if (!exceptions.isEmpty()) {
-        throw exceptions.get(0);
-      }
-    }
-  }
-}
